Strafing improved.
We need to stop using the teleopContinuous() method, as there's another
way to handle it that Mr. Holincheck described. That will be added
shortly, as it shouldn't be hard to implement. 

If I un-comment out the talons declared in the MotorSubsystem class, I
get an unhandeled exception PWM already allocated to X port where X is
the parameter I use when constructing the talons.